pyproject.toml
[project]
name = "deckaesthetics"
version = "0.1.0"
description = "Analyze MTG decklists for available frame/border/aesthetic printings"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "duckdb>=1.1",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "pyyaml>=6.0",
    "pydantic>=2.9",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.7",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["backend"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
